Move data to "Common" domain
