+++ Packages/App/Sources/CommonDataPreferences/Domain/CDExcludedEndpoint.swift
