+++ Packages/App/Sources/CommonDataPreferences/Domain/CDFavoriteServer.swift
