+++ Packages/App/Sources/CommonDataProviders/Strategy/CDProviderRepositoryV3.swift
@@ -23,7 +23,7 @@
 //  along with Passepartout.  If not, see <http://www.gnu.org/licenses/>.
 //
 
-import AppData
+import CommonData
 import CommonLibrary
 import CommonUtils
 import CoreData
